switch met library with separated out one. fix #534
