+++ Godeps/_workspace/src/github.com/raintank/met/dogstatsd/init.go
@@ -1,5 +1,4 @@
 // a metrics class that uses dogstatsd on the backend
-// at some point, we could/might try to unite this with the metrics package
 
 // note that on creation, we automatically send a default value so that:
 // * influxdb doesn't complain when queried for series that don't exist yet, which breaks graphs in grafana
