+++ Godeps/_workspace/src/github.com/raintank/met/interfaces.go
@@ -1,4 +1,4 @@
-package metric
+package met
 
 import "time"
