+++ Godeps/_workspace/src/github.com/raintank/met/statsd/gauge.go
@@ -1,10 +1,11 @@
-package dogstatsd
+package statsd
 
 import (
 	"sync"
 	"time"
+
+	"github.com/raintank/met"
 )
-import "github.com/grafana/grafana/pkg/metric"
 
 type Gauge struct {
 	key string
@@ -13,7 +14,7 @@ type Gauge struct {
 	backend Backend
 }
 
-func (b Backend) NewGauge(key string, val int64) metric.Gauge {
+func (b Backend) NewGauge(key string, val int64) met.Gauge {
 	g := Gauge{
 		key:     key,
 		backend: b,
@@ -21,7 +22,7 @@ func (b Backend) NewGauge(key string, val int64) metric.Gauge {
 	go func() {
 		for {
 			g.Lock()
-			g.backend.client.Gauge(g.key, float64(g.val), []string{}, 1)
+			g.backend.client.Gauge(g.key, int(g.val))
 			g.Unlock()
 			time.Sleep(time.Duration(1) * time.Second)
 		}
@@ -31,21 +32,21 @@ func (b Backend) NewGauge(key string, val int64) metric.Gauge {
 
 func (g *Gauge) Value(val int64) {
 	g.Lock()
-	defer g.Unlock()
 	g.val = val
-	g.backend.client.Gauge(g.key, float64(g.val), []string{}, 1)
+	g.Unlock()
+	g.backend.client.Gauge(g.key, int(val))
 }
 
 func (g *Gauge) Inc(val int64) {
 	g.Lock()
 	defer g.Unlock()
 	g.val += val
-	g.backend.client.Gauge(g.key, float64(g.val), []string{}, 1)
+	g.backend.client.Gauge(g.key, int(g.val))
 }
 
 func (g *Gauge) Dec(val int64) {
 	g.Lock()
 	defer g.Unlock()
 	g.val -= val
-	g.backend.client.Gauge(g.key, float64(g.val), []string{}, 1)
+	g.backend.client.Gauge(g.key, int(g.val))
 }
