+++ pkg/services/collectoreventpublisher/collectoreventpublisher.go
@@ -8,9 +8,9 @@ import (
 	"time"
 
 	"github.com/grafana/grafana/pkg/log"
-	met "github.com/grafana/grafana/pkg/metric"
 	"github.com/grafana/grafana/pkg/setting"
 	"github.com/nsqio/go-nsq"
+	"github.com/raintank/met"
 	"github.com/raintank/raintank-metric/schema"
 )
